backend/requirements.txt
# Web Framework
flask==3.1.0
flask-cors==5.0.0

# Google Gemini AI
google-generativeai==0.8.3

# Environment Variables
python-dotenv==1.0.1

# NLP & Text Processing
spacy==3.8.2
# Download model after install: python -m spacy download en_core_web_sm

# Data Analysis
pandas==2.2.3
numpy==2.2.1

# Visualization
matplotlib==3.10.0
seaborn==0.13.2

# Emoji Handling
emoji==2.14.0

# HTTP Requests (for testing)
requests==2.32.3